dian-proxy/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
playwright==1.49.1
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.10.4
python-dotenv==1.0.1
numpy>=1.26.0
